refractor some code

+++ controllers/api/comment-routes.js
@@ -2,47 +2,49 @@ const router = require('express').Router();
 const { Comment } = require('../../models');
 
 router.get('/', (req, res) => {
-    Comment.findAll({
-        attributes: [
-            'id',
-            'post_url',
-            'title',
-            'created_at',
-        ],
-        order: [['created_at', 'DESC']],
-    })
-});
-
-router.post('/', (req, res) => {
-    Comment.create({
-        comment_text: req.body.comment_text,
-        user_id: req.body.user_id,
-        post_id: req.body.post_id
-    })
+    Comment.findAll()
         .then(dbCommentData => res.json(dbCommentData))
         .catch(err => {
             console.log(err);
-            res.status(400).json(err);
-        });
+            res.status(500).json(err)
+        })
 });
 
-router.delete('/', (req, res) => {
-    Comment.destroy({
-        where: {
-            id: req.body.post_id
-        }
-    })
-        .then(dbCommentData => {
-            if (!dbCommentData) {
-                res.status(404).json({ message: 'No post found with this id' });
-                return;
-            }
-            res.json(dbCommentData);
+router.post('/', (req, res) => {
+    //check the session
+    if (req.session) {
+        Comment.create({
+            comment_text: req.body.comment_text,
+            user_id: req.body.user_id,
+            post_id: req.body.post_id
         })
-        .catch(err => {
-            console.log(err);
-            res.status(500).json(err);
-        });
+            .then(dbCommentData => res.json(dbCommentData))
+            .catch(err => {
+                console.log(err);
+                res.status(400).json(err);
+            });
+    }
+});
+
+router.delete('/:id', (req, res) => {
+    if (req.session) {
+        Comment.destroy({
+            where: {
+                id: req.params.id
+            }
+        })
+            .then(dbCommentData => {
+                if (!dbCommentData) {
+                    res.status(404).json({ message: 'No comment found with this id' });
+                    return;
+                }
+                res.json(dbCommentData);
+            })
+            .catch(err => {
+                console.log(err);
+                res.status(500).json(err);
+            });
+    }
 });
 
 module.exports = router;
